correct the graphic buffer size allocated when rotating the device
allocateBuffers() is called with H x W when a device is rotated
to landscape orientation. However, the EGL implementation
requests buffers with pre-rotated size, W x H as it is reacting
to the transform hint.
So if transform hint is set, we need to swap the requested
width and height to avoid the buffers being freed and
reallocated when dequeueBuffers() gets called later.

Test: perform systrace during rotating a marlin device.